[FMD-26] Remove old methods/code/tests
